organize schemas in namespaces
