+++ libs/core/src/framework/index.ts
@@ -1,5 +1,4 @@
 export * from './command';
 export * from './event';
-export * from './pagination';
 export * from './query';
 export * from './types';
