+++ libs/core/src/index.ts
@@ -2,8 +2,7 @@ export * as commonProtocol from './commonProtocol';
 export * from './errors';
 export * from './framework';
 export * from './ports';
-export * from './requirements-types';
-export * from './schemas';
+export * as schemas from './schemas';
 export * from './shared-types';
 export * from './types';
 export * from './utils';
