+++ libs/core/src/schemas/commands/address.schemas.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod';
+
+export const TodoAddressCommands = z.object({});
