+++ libs/core/src/schemas/commands/chainNode.schemas.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod';
+
+export const TodoChainNodeCommands = z.object({});
