+++ libs/core/src/schemas/commands/comment.schemas.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod';
+
+export const TodoCommentCommands = z.object({});
