+++ libs/core/src/schemas/commands/thread.schemas.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod';
+
+export const TodoThreadCommands = z.object({});
