+++ libs/core/src/schemas/commands/user.schemas.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod';
+
+export const TodoUserCommands = z.object({});
