+++ libs/core/src/schemas/events.schemas.ts
@@ -1,16 +1,12 @@
 import { z } from 'zod';
 
-export const schemas = {
-  ThreadCreated: z.object({ thread: z.string() }),
-  CommentCreated: z.object({ comment: z.string() }),
-  GroupCreated: z.object({
-    groupId: z.string(),
-    userId: z.string(),
-  }),
-  CommunityCreated: z.object({
-    communityId: z.string(),
-    userId: z.string(),
-  }),
-};
-
-export type Events = keyof typeof schemas;
+export const ThreadCreated = z.object({ thread: z.string() });
+export const CommentCreated = z.object({ comment: z.string() });
+export const GroupCreated = z.object({
+  groupId: z.string(),
+  userId: z.string(),
+});
+export const CommunityCreated = z.object({
+  communityId: z.string(),
+  userId: z.string(),
+});
