--- libs/core/src/schemas/notification.schemas.ts
@@ -1,8 +1,0 @@
-import z from 'zod';
-
-export const NotificationCategory = z.object({
-  name: z.string().max(255),
-  description: z.string(),
-  created_at: z.date(),
-  updated_at: z.date(),
-});
