--- libs/core/src/schemas/thread.schemas.ts
@@ -1,7 +1,0 @@
-import z from 'zod';
-import { Address } from './address.schemas';
-
-// TODO: move attributes from models
-export const Thread = z.object({
-  Address: Address.optional(),
-});
