+++ libs/model/src/community/CreateCommunity.command.ts
@@ -1,12 +1,12 @@
 import type { CommandMetadata } from '@hicommonwealth/core';
-import { community } from '@hicommonwealth/core';
+import { schemas } from '@hicommonwealth/core';
 import { models } from '../database';
 import { mustNotExist } from '../middleware/guards';
 
 export const CreateCommunity = (): CommandMetadata<
-  typeof community.CreateCommunity
+  typeof schemas.commands.CreateCommunity
 > => ({
-  schemas: community.CreateCommunity,
+  schemas: schemas.commands.CreateCommunity,
   auth: [],
   body: async ({ id, payload }) => {
     console.log(payload); // TODO: remove
