+++ libs/model/src/community/UpdateCommunity.command.ts
@@ -1,13 +1,13 @@
-import { CommandMetadata, community } from '@hicommonwealth/core';
+import { CommandMetadata, schemas } from '@hicommonwealth/core';
 import { models } from '../database';
 import { isCommunityAdmin } from '../middleware';
 import { mustExist } from '../middleware/guards';
 import { commonProtocol } from '../services';
 
 export const UpdateCommunity = (): CommandMetadata<
-  typeof community.UpdateCommunity
+  typeof schemas.commands.UpdateCommunity
 > => ({
-  schemas: community.UpdateCommunity,
+  schemas: schemas.commands.UpdateCommunity,
   auth: [isCommunityAdmin],
   body: async ({ id, payload }) => {
     const community = await models.Community.findOne({ where: { id } });
