+++ libs/model/src/reaction/CreateReaction.command.ts
@@ -1,4 +1,4 @@
-import { todo, type CommandMetadata } from '@hicommonwealth/core';
+import { schemas, type CommandMetadata } from '@hicommonwealth/core';
 import { z } from 'zod';
 import { models } from '../database';
 import { mustNotExist } from '../middleware/guards';
@@ -7,8 +7,10 @@ export const schema = z.object({
   content: z.string(),
 });
 
-export const CreateReaction = (): CommandMetadata<typeof todo.CreateTodo> => ({
-  schemas: todo.CreateTodo,
+export const CreateReaction = (): CommandMetadata<
+  typeof schemas.commands.CreateTodo
+> => ({
+  schemas: schemas.commands.CreateTodo,
   auth: [],
   body: async ({ id, payload }) => {
     const reaction = await models.Reaction.findOne({ where: { id } });
