+++ libs/model/src/test/seed.ts
@@ -1,4 +1,5 @@
 import { generateMock } from '@anatine/zod-mock';
+import { schemas } from '@hicommonwealth/core';
 import { Model, ModelStatic } from 'sequelize';
 import z from 'zod';
 
@@ -17,6 +18,8 @@ type GeneratedProp = typeof GENERATED_PROPS[number];
 // incrementing seed number for each call to the `seed` function
 // for deterministic randomness when the order of calls is preserved
 let seedNum = 1;
+
+type E = keyof typeof schemas;
 
 export type SchemaWithModel<T extends z.AnyZodObject> = {
   schema: T;
