+++ libs/model/src/user/CreateUser.command.ts
@@ -1,9 +1,11 @@
-import { todo, type CommandMetadata } from '@hicommonwealth/core';
+import { schemas, type CommandMetadata } from '@hicommonwealth/core';
 import { models } from '../database';
 import { mustNotExist } from '../middleware/guards';
 
-export const CreateUser = (): CommandMetadata<typeof todo.CreateTodo> => ({
-  schemas: todo.CreateTodo,
+export const CreateUser = (): CommandMetadata<
+  typeof schemas.commands.CreateTodo
+> => ({
+  schemas: schemas.commands.CreateTodo,
   auth: [],
   body: async ({ id, payload }) => {
     const user = await models.User.findOne({ where: { id } });
