+++ packages/commonwealth/server/routes/ddd/integrations.ts
@@ -1,11 +1,11 @@
 import { trpc } from '@hicommonwealth/adapters';
-import { analytics, events, type PolicyMetadata } from '@hicommonwealth/core';
+import { analytics, schemas, type PolicyMetadata } from '@hicommonwealth/core';
 import { MixpanelCommunityInteractionEvent } from '../../../shared/analytics/types';
 
 const inputs = {
-  GroupCreated: events.schemas.GroupCreated,
-  ThreadCreated: events.schemas.ThreadCreated,
-  CommentCreated: events.schemas.CommentCreated,
+  GroupCreated: schemas.events.GroupCreated,
+  ThreadCreated: schemas.events.ThreadCreated,
+  CommentCreated: schemas.events.CommentCreated,
 };
 
 const Analytics: () => PolicyMetadata<typeof inputs> = () => ({
